requirements.txt
Django==4.2.25
django-markdownfield==0.11.0